zenith_rmdir function added with recursive capabilitiy.
By now only zenith_rmdir can make use of the extended flag and the
extended address present inside a general node. With those option to
move around i found necessary to add 2 function and chose a work
operation of function and subfunction.
For simplicity the function layer serve the standard rmdir function,
inside it afted the sandard startup steps the __rmdir function would be
called out to analize the content of the folder and find some maches
with the provided name. If so it return a state to the main function and
finish the operation, if not it will excange the cached folder address
with the extended one ( if present ) and then it will call itself to
analize the content of the extended folder, and the process will
continue until some error or matches will be found.