discontinuing virtual drive support ( for now )
For the previous commits i introduce some difference inside all function
to integrate the virtual drive with the functions of zenith. Well ..
this is not a good way to introduce the "virtualization" for the simple
fact that all funtions will be differents in different points, and this
is fine until you need to debug thing because you are not 100% sure that
what you have write for the virtual drive will function perfectly in the
direct drive support ( bare metal support ). All function will just be
differend even if you try to build it as similar you can.

So i decided to drop the current implementation of the virtual drive and
work around a more "elegant" solution by indroduce a new layer inside
the zenith interface layer that will operate to save data in ram if
zenith is compiled with VIRTUAL_DRIVE flag enabled. So there is no such
defferents in the main function, they will work just like on a 8bit
computer, the only visible changes would be around the interface level,
the deepest so far.